% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svdvis-package.R
\docType{package}
\name{svdvis-package}
\alias{svdvis}
\alias{svdvis-package}
\title{Singular Value Decomposition Visualization}
\description{
Visualize singular value decompositions (SVD), principal component analysis (PCA), factor analysis (FA) and related methods.
}
\details{
\tabular{ll}{
Package: \tab svdvis\cr
Type: \tab Package\cr
Version: \tab 1\cr
License: \tab GPL-2\cr
Depends: \tab ggplot2, GGally, gridExtra, scales, reshape2, RColorBrewer, grDevices\cr
}

Providing convenient visualization functions for paired scatterplots, heat maps, parallel coordinates, and radial coordinates.
The package is developed for the outputs from singular value decompositions (SVD), principal component analysis (PCA), factor analysis (FA), and related methods.
Particularly, it focuses on exploring and visualizing multidimensional vectors (singular vectors, principal components, or factors) from dimension reduction techniques or latent variable models.
}
\author{
Neo Christopher Chung \email{nchchung@gmail.com}
}

